refactor(cli): centralize duplicated prompt and readConfig functions
Extracted `prompt` and `readConfig` functions, along with common path constants, into a new `cli/src/utils.ts` module to improve maintainability and follow DRY principles.

Changes:
- Created `cli/src/utils.ts` with shared functions and constants.
- Refactored `bootstrap.ts`, `env.ts`, `service.ts`, and `skills.ts` to use the new utility module.
- Cleaned up unused imports in the refactored files.
